+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Support attributes added to the config. Allows an implementation to add driver
   options that will be included if a connection has to be reconnected.
   Add supported driver options as an array on the `attributes` config key.
+### Changed
+- Stop using a prepared statement when setting charset and timezone.
+  `SET NAMES` isn't supported when using native prepared statements.
 
 ## [2.0.0] - 2021-08-07
 ### Added